refactor: :recycle: Change AlertType in handleLendbtn
Change AlertType in handleLendbtn from CONFIRMATION to INFORMATION to avoid misunderstandings: when a user lend a movie, the confirmation alert is shown, but the alert has a "ok" and a "cancel" button. This can be misleading to the user since clicking "cancel" in the alert will not cancel the action.