refactor: :recycle: Reconfigure the MoviePageController.java to handle Movies.json and rewrite javadoc comments
Reconfigured the MoviePageController.java to handle Movies.json instead of Movies.csv by importing the classes: MovieDeserializer, MovieSerializer and MovieManager. Declared variables with the corresponding types to later create objects used for datahandling in the initialize method. Reconfigured the handleLendbtn and handleReturnbtn methods to call the methods from MovieSerializer and MovieDeserializer. Added a call to the serialize method in MovieSerializer to update the movie status to lent (true) in the Movies.json file. Rewrote javadoc comments.

Co-authored-by: Helle Hornes Fystro