feat: :sparkles: Create and implement temporary solution to a REST controller
Created and implemented temporary solution to a REST controller called "MovieLibraryController.java". This is not working due to that it is impossible to reference to the movies.json file through a File object when the application is already packaged into a .jar. The purpose of commiting and merging this in is to test alternative approaches and testing.

Co-authored-by: Sylvia Yung